requirements.txt
ultralytics>=8.1.0
pyyaml>=6.0
torch>=2.0.0
numpy>=1.23
opencv-python>=4.7